fix(providers): address OpenRouter STT review feedback
- Extract the OpenRouter transcription handler into
  open-sse/handlers/openrouterTranscription.ts to keep audioTranscription.ts
  under the file-size cap.
- Qualify the STT card's submitted model id with the connection's provider
  prefix so OpenRouter models route to OpenRouter rather than the model's own
  vendor (the transcription route resolves the provider from the leading
  segment of the model id).
- Coerce temperature to a number and forward timestamp_granularities on the
  JSON input_audio payload; match the base MIME type when resolving the audio
  format so codec parameters (e.g. audio/webm;codecs=opus) do not fall back to wav.
- Split the OpenRouter cases into tests/unit/audio-transcription-openrouter.test.ts
  and add coverage for temperature, timestamp granularities, MIME codec params,
  and qualified-id routing.
