+++ tests/unit/audio-transcription-handler.test.ts
@@ -821,64 +821,3 @@ test("handleAudioTranscription rejects Gladia jobs missing a result_url", async 
     globalThis.fetch = originalFetch;
   }
 });
-
-const OPENROUTER_TRANSCRIPTION_MODELS = [
-  "deepgram/nova-3",
-  "microsoft/mai-transcribe-1.5",
-  "nvidia/parakeet-tdt-0.6b-v3",
-  "mistralai/voxtral-mini-transcribe",
-  "qwen/qwen3-asr-flash-2026-02-10",
-  "google/chirp-3",
-  "openai/gpt-4o-mini-transcribe",
-  "openai/whisper-large-v3",
-  "openai/whisper-large-v3-turbo",
-  "openai/whisper-1",
-  "openai/gpt-4o-transcribe",
-];
-
-test("OpenRouter transcription registry accepts every supported nested model id", async () => {
-  const { parseTranscriptionModel } = await import("../../open-sse/config/audioRegistry.ts");
-  for (const model of OPENROUTER_TRANSCRIPTION_MODELS) {
-    assert.deepEqual(parseTranscriptionModel(`openrouter/${model}`), {
-      provider: "openrouter",
-      model,
-    });
-  }
-});
-
-test("OpenRouter transcription converts multipart audio to dedicated STT input_audio", async () => {
-  const originalFetch = globalThis.fetch;
-  let capturedUrl = "";
-  let capturedInit;
-  globalThis.fetch = async (url, init) => {
-    capturedUrl = String(url);
-    capturedInit = init;
-    return Response.json({ text: "hello from MAI" });
-  };
-
-  try {
-    const form = new FormData();
-    form.set("file", buildFile([1, 2, 3], "sample.flac", "audio/flac"));
-    form.set("model", "openrouter/microsoft/mai-transcribe-1.5");
-    form.set("language", "pt");
-
-    const response = await handleAudioTranscription({
-      formData: form,
-      credentials: { apiKey: "or-test-key" },
-    });
-
-    assert.equal(response.status, 200);
-    assert.deepEqual(await response.json(), { text: "hello from MAI" });
-    assert.equal(capturedUrl, "https://openrouter.ai/api/v1/audio/transcriptions");
-    const headers = capturedInit?.headers as Record<string, string>;
-    assert.equal(headers.Authorization, "Bearer or-test-key");
-    assert.equal(headers["Content-Type"], "application/json");
-    const body = JSON.parse(String(capturedInit?.body));
-    assert.equal(body.model, "microsoft/mai-transcribe-1.5");
-    assert.equal(body.input_audio.format, "flac");
-    assert.equal(body.input_audio.data, "AQID");
-    assert.equal(body.language, "pt");
-  } finally {
-    globalThis.fetch = originalFetch;
-  }
-});
